Fix parsing by using libxml
* smbc.el (smbc-parse-html): Replace regexps and string manipulation
with libxml and DOM manipulation.  Because the libxml-parse-html-region
function requires a buffer as input, replace the string argument to
smbc-parse-html with a buffer argument.  Because SMBC no longer
consistently uses sequentially integers for image ids, parse the links
to the previous and next comics and store using the smbc-previous-url
and smbc-next-url variables.
(smbc-retrieve): New function that retrieves the given URL or path on
smbc-comics.com and returns a buffer containing the content.
(smbc-get-previous, smbc-get-next): Use smbc-retrieve and the
smbc-previous-url or smbc-next-url variable, respectively, or print
a message indicating if the variable is null.
(smbc-get-image-data, smbc-get-index-page): Use smbc-retrieve.
(smbc-get-page-given-id, smbc-chomp): Delete.